Capture allocation events for child jobs and log new allocations
This change refactors the monitor into a struct that tracks
allocations that have previously been seen so that new
allocations may be picked up for logging.

It also naively checks for subjobs of batch jobs so that
each run may be captured.

There's a potential timing issue if an allocation takes some
time to be placed.